Added rudimentary input via text commands
Added a textBox and a Button component to the form. If the parseButton
is clicked the TimeLine is cleared, the text in the scriptBox is parsed
to TimeLineItem's and then added to the TimeLine.
It allows for input via text and supports some shorthand notations for
common inputs (like shoryuken or quarter circle).

Syntax:
Every line is interpreted as one TimeLineItem.
Simultaneous commands are input via '.' (eg HP.MP = Press HP MP)
HoldItems are input via '+' (eg +D = Hold DN)
ReleaseItems are input via '-' (eg -D = Release DN)
WaitFrameItems are input via 'W' followed by the number of frames to
wait (eg W100 = Wait 100 Frames)
Every other input is interpreted as a PressItem

It should ignore every wrong or unrecognized input.