Control for the time line input now holds the settings for sound and wait frames. Keeping everything together will make binding easier.
This is now depreceated - these settings are found in the input item Control

Refactored the timelineitem into TimeLineItemModel and setup the ground work for using a state struct rather than an array of enums.

Explicit function in InputResolver to check if player is on the left

Reorganised the API calls in SF4Control to be at the top of the file and added regions to make file more reader friendly

The huge change that makes this branch meaningful. This changed the heirarchy of the InputItem so that there is no WaitItem and that a wait is simply an input without any inputs to send. This will make time line construction simpler for users and easier to handle in the long run for the developers.

Include the InputSimulator dll in the Debug folder.

The Great Refactor - the refactor to end all refactors - renaming the timelineitems to be appropriate to the new model and reforming the way we construct inputs items on the time line.

Cosmetic changes, set sizes on the Controller Interface and removed the TimeLineItemDetails control form the main form (no need to have it if the Wait Item is now functionally the same as an Input Item)

Adding in the properties and functionallity for InputStates to be used in the TimeLineItemViewModel so that we can set the directions, inputs and extra settings of a time line item from the ViewModel itself, this will enable binding it to the source when in the View and make adjusting items on the time line easier.

Description accessible within new TimeLineItemModels

Awesome new bindings for the front end in WPF, Up, Down, Left, Right and all Punch/Kick buttons now save state in the time line. The Wait Frames has been made public too so that they can be set as an observable object. You can now click an item and see it's button layout and wait frame setup without needing to "translate" the Token string in the list.

The tool is now slimmer, the list is on top of the controller interface and the Press, Hold, Release buttons are in place.

Hold/Release will probably be changed and incorporated into the toggle buttons (toggle once for press, twice for hold and create a new TimeLineItem where the button is neutral to signify a release)